move Tape back to the toplevel module of tape
